Move original dashboard to /workouts route in order to change dashboard responsibility
